+++ http-interface/HttpServiceConfig.cs
@@ -1,0 +1,27 @@
+// Copyright 2025 Keyfactor
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Net.Http.Headers;
+using System.Threading.Tasks;
+
+namespace Keyfactor.Extensions.Utilities.HttpInterface
+{
+    public class HttpServiceConfig
+    {
+        public string Accept { get; set; } = "application/json";
+        public string ContentType { get; set; } = "application/json";
+        public Func<Task<AuthenticationHeaderValue>> AuthorizationDelegate { get; set; }
+    }
+}
